Tracking down an issue with n_jobs>0
Former-commit-id: 5730403fd77d32dd22de9850ee761a601e8d5ce4

+++ tests/test_torch_optimization_util.py
@@ -8,7 +8,7 @@ from torch.func import grad_and_value, hessian
 def test_lm_basic():
     n = 10
     p = 11
-    x0 = torch.ones((n, p))
+    x0 = torch.ones((n, p)).double()
 
     # just checking we can minimize the norm
     def obj(z):
